PasswordReset.java
// PasswordReset.java
public class PasswordReset {
    public boolean resetPassword(String username, String newPassword) {
        // Validate username and newPassword
        if (isValidUsername(username) && isValidPassword(newPassword)) {
            // Reset the password in the data store
            return true;
        } else {
            return false;
        }
    }

    private boolean isValidUsername(String username) {
        return username != null && !username.isEmpty();
    }

    private boolean isValidPassword(String password) {
        // Implement your password validation logic here
        // For demonstration purposes, let's assume a password must be at least 8 characters long
        return password != null && password.length() >= 8;
    }
}
